tests/wpt/web-platform-tests/layout-instability/content-visibility-auto-offscreen.html

<!DOCTYPE html>
<title>Layout Instability: off-screen content-visibility:auto content</title>
<link rel="help" href="https://wicg.github.io/layout-instability/" />
<style>
  #auto {
    content-visibility: auto;
    contain-intrinsic-size: 1px;
    width: 100px;
  }
</style>
<div class=auto>
  <div style="width: 100px; height: 100px"></div>
</div>
<div class=auto style="position: relative; top: 100000px">
  <div style="width: 100px; height: 100px"></div>
</div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/util.js"></script>
<script>
promise_test(async () => {
  const watcher = new ScoreWatcher;

  // Wait for the initial render to complete.
  await waitForAnimationFrames(2);

  window.scrollTo(0, 100000 + 100);
  await waitForAnimationFrames(2);

  assert_equals(watcher.score, 0);

  window.scrollTo(0, 0);
  await waitForAnimationFrames(2);

  assert_equals(watcher.score, 0);
}, 'off-screen content-visibility:auto');

</script>
